separate fellowships related bits to its own js file
